rename console directory to commands
